requirements.txt
Django==4.2.26
djangorestframework==3.14.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0